Refactor add into import
